backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.9.0
executing==2.0.1
Flask==3.0.3
Flask-Cors==4.0.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.64.0
groq==0.9.0
grpcio==1.66.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.1
idna==3.8
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.2.35
langchain-google-genai==1.0.10
langsmith==0.1.104
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
orjson==3.10.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==8.5.0
textwrap3==0.9.2
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
